[IGNORE] login device in Identity Service
Summary:
NOTE: This code will not be shipped.

I need this code to properly test creating `olm` sessions.
Publishing at begging of stack to make it easier for reviewers, and maybe it could be helpful to also test it in the future.

Depends on D10369

Test Plan: Connection with Tunnelbroker should initialized

Reviewers: marcin, varun, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10327