[comm-lib] Add BlobInfo::from_bytes
Summary:
Small refactor, that makes the code a bit cleaner. This diff also moves the required dependencies (that will also be used in the next diff). The `ReportContent` struct that uses this function will also be moved to `comm-lib`, as it will be useful for the backup service (keeping logs in either blob or dynamodb).

Test Plan:
`cargo check` in report service

Reviewers: kamil, bartek

Subscribers: