[comm-lib] Function for calculating dynamodb item size
Summary:
Now that we will have two places where we need to calculate the size of an item in dynamodb, it would be good to introduce an abstraction that will do it for us. This diff introduces a function that will take a `AttributeMap` and return size of the item. The specific values were taken from [aws docs](https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/CapacityUnitCalculations.html), and [this blogpost](https://zaccharles.medium.com/calculating-a-dynamodb-items-size-and-consumed-capacity-d1728942eb7c).

Differences from the current method (in terms of the final size):
- this method doesn't count length of optional attribute names if they don't exist in the final item
- lest error-prone -> e.g. the current implementation skipped value of `creation_time` field during size calculations

A problem with this approach is unnecessary allocations when creating the `AttributeMap` that will be then discarded. There's a chance that Rust will optimize it away, but I haven't tested it. In any case the "correct" solution would be something with either `derive` macros or custom serializers with `serde`. But the code in this diff is very localized so it will be easy to replace if it turned out to be a performance problem in the future.

Depends on D10402

Test Plan:
`cargo check` in report service.
I run the current calculations and the new function, and for this item:
```
ReportItem {
  id: ReportID::default(),
  user_id: "user_id".to_string(),
  report_type: ReportType::ErrorReport,
  platform: ReportPlatform::MacOS,
  creation_time: Utc::now(),
  content: BlobOrDBContent::Blob(BlobInfo::from_bytes(b"data")),
  encryption_key: None,
};
```
The current soluton returned `230` and the new one returned `259`. The slight difference can be attributed to counting the length of `creation_time` field, and skipping the `encryption_key` attribute name

Reviewers: bartek, kamil

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10405