[web] introduce calculateNavigationSidebarTooltipSize
Summary: Part of the process when we create tooltips is that we need to calculate the size of the tooltip. This diff introduces the function that will handle calculating the size of the tooltip.

Test Plan: Logged out results of this function also inspected the tooltip element (w/o passing the `style` prop into the tooltip) and confirmed that I was getting the correct size

Reviewers:

Subscribers: