[web] introduce calculateNavigationSidebarTooltipSize
Summary:
Part of the process when we create tooltips is that we need to calculate the size of the tooltip. This diff introduces the function that will handle calculating the size of the tooltip.

Part of https://linear.app/comm/issue/ENG-5948/hovering-over-every-clickable-element-in-this-sidebar-will-have-a

Test Plan:
Logged out results of this function also inspected the tooltip element (w/o passing the `style` prop into the tooltip) and confirmed that I was getting the correct size

{F954720}

{F954719}

Reviewers: atul, rohan, kamil, ashoat, tomek

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10357