[web] introduce community list item component
Summary:
This diff introduces the community list item component. This component will be used in a community list component that I will introduce in a future diff.

For context on what this should look like here is a screenshot from the figma:
{F954850}

Part of https://linear.app/comm/issue/ENG-5947/list-of-all-top-level-communites-communities

Depends on D10358

Test Plan:
Please see the screenshot below

{F954845}

Reviewers: atul, rohan, kamil

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10359