Check against 'workerRegistration.pushManager' before creating push subscription on web
Summary:
This differential solves https://linear.app/comm/issue/ENG-6006/. Even though web documentation doesn't allow `pushManager` property of `workerRegistration`
to be nullish, there are some recources that suggest it could actually be the case so we add additional check to protect against the issue.

Test Plan:
Check that web notifications are working properly. We can't reproduce the issue as it is probably some race condition so we can't directly test this change. However
this diff should protect against it since we won't let access `pushManager` properties in case it is nullish.

Reviewers: michal, varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10351