[CommCoreModule] implement decrypting content with `olm` session
Summary:
Implemented method to decrypt content.

Test Plan:
This code is really hard to test as it's impossible to verify partial results.
I generated some keys and values, verified them with `olm-utils.test.js`, and checked if they succeeded/failed when calling the method introduced here in the same cases as JS tests.
Full functionality tested later in the stack.

Reviewers: marcin, varun, michal