[native] add button to test send encrypted content between devices
Summary:
Sending encrypted messages to different devices.
As a follow-up, we can consider adding some alerts to make it easier to test on prod but still, at current state it could only be tested on the dev end.

Depends on D10382

Test Plan:
This is the last diff on which I tested everything end-to-end
1. Two devices on the same account
2. Connected to staging Identity and Tunnelbroker
3. Created a session (D10382)
4. Verified that session is created on both devices
5. Use this to send encrypted message to device (both ways)
6. Look at logs and test if message is decrypted

Reviewers: marcin, varun, michal

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10383