[terraform] Make Tunnelbroker discoverable for Identity
Summary:
Noticed that there are multiple error logs on staging that Identity wasn't able to ask the device for more one-time keys because it couldn't reach Tunnelbroker.
This diff exposes Tunnelbroker gRPC port to the ECS Service Connect and updates the Identity configuration to use it.

Test Plan:
Tested on staging, @kamil confirmed that he was able to get the "refresh request" message:
{F980383}

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek, kamil

Differential Revision: https://phab.comm.dev/D10367