[lib/web/native] Show and search resolved ENS names in @ mentioning toolip
Summary: In this diff, we modify the @ mentioning tooltip for users to display resolved ENS names and also to allow users to type out ENS names and still get a match. For example, for a user `imbatman.eth`, if I type @ followed by '0x', I should see `imbatman.eth`. I should also be able to see this if I type '@im' now.

Because of these changes, we're able to remove unused code in `user-selectors`.

Test Plan: Please see the videos below for both `web` and `native`

Reviewers: atul, ginsu, inka

Subscribers: