[services] [5/n] prefix search websocket server
Summary:
This diff implements a Websocket server on the identity service that queries users on the Opensearch domain with prefixes. Currently, these diffs implements just the upgrade and prefix searching. Diffs for authentication are pending testing.

The websocket server is initialized on port 50057 on the identity server.

Depends on D9878

Test Plan: Tested on staging with a custom image. Upgrade successful and prefix search returns a list of user_id, username pairs.

Reviewers: bartek, varun, kamil, O1 Nix

Reviewed By: varun, O1 Nix

Subscribers: kamil, ashoat, tomek

Differential Revision: https://phab.comm.dev/D10392