[native] remove android feature
Summary: this feature is unnecessary. we can just check the target_os as done in D10263, and set DEVICE_TYPE accordingly.

Test Plan: ran the android emulator and confirmed that DEVICE_TYPE was set to the Android variant by inspecting the version interceptor on identity service. On iOS simulator, DEVICE_TYPE was set to the iOS variant.

Reviewers: bartek, michal!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10394