[web] don't log out user when processing draft store ops fails
Summary:
Each time one character is typed operation with updated text is processed, so there is a high possibility that if something fails, those are draft operations.

When e.g. thread store operation fails we end with a corrupted state which could cause many issues, so it's safe to delete the database and log out the user.

When the draft operation fails at the worst case we lose the last character typed (if we stopped typing after that last character on which there was a failure), in other cases this does not influence app logic, so I think we can avoid logging to mitigate the user's frustration and at the same time keep alert with logging which could help debugging issues.

Test Plan: Throw when updating the draft and check if only the alert is shown without logging the user out

Reviewers: tomek, michal

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D10395