[identity] Store device code version and login date
Summary:
For primary device "migration", we need to store device code version along with the timestam at which access token was generated. This is required for [[ https://linear.app/comm/issue/ENG-5841/reorder-devices-upon-loginregister | ENG-5841 ]].

Depends on D10397

Test Plan:
Registered a new user and verified that the device code version and login date are stored in the database.
Logged in as that user but with a new device, made sure the data is stored for the new device too.

Reviewers: michal, varun

Reviewed By: varun

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10398