[lib] Add function for filtering out calendar filters related to one keyserver
Summary:
issue: [[ https://linear.app/comm/issue/ENG-5810/create-a-function-to-set-default-calendar-filters-to-a-subset-of | ENG-5810 ]]
I added a helper function `removeKeyserverThreadIDsFromFilterList`, which removes thread ids of theads reladted to given keyservers from calendar filters.
This function is very similar to previously existing `removeDeletedThreadIDsFromFilterList`, so I extracted the common code

Test Plan: ran the provided tests. The tests for `removeDeletedThreadIDsFromFilterList` are not very exhaustive, but they check that the functinon didn't stop working as a result of my changes.

Reviewers: michal, kamil, ginsu

Reviewed By: michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10425