[lib] Move `socketPreferredEndpoints` to `httpPreferredEndpoints`
Summary:
Context is here: https://linear.app/comm/issue/ENG-6134/switch-endpoints-from-socket-to-http

Intentionally leaving `socketPreferredEndpoints` and accompanying functionality around for the future if we ever want to prefer socket for any of these endpoints. Figure if we have the machinery in place, we can keep it around even if `socketPreferredEndpoints` is empty for the time being. Let me know if this is an incorrect assumption and I can rip out `socketPreferredEndpoints`, `SocketPreferredEndpoint`, etc.

Test Plan:
Went through `native` app connected to prod `keyserver` and tried to "hit" as many of the endpoints as possible (eg sent messages, created calendar entries, sent image message, created threads, etc) and things continued to work as expected. However, it's worth noting that I haven't yet seen the "disconnected issue" on build 302 yet, so I'm not able to confirm whether this reduces timeouts, etc.

Reviewers:
ashoat, ginsu, tomek, rohan