[lib] Remove cookie on logout
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6260/set-cookie-to-null-on-logout | ENG-6260 ]]
If the user tries to log out, but the keyserver doesn't respond, we timeout and dispatch a logout success anyway.
But if the keyserver is down, it won't send a session change. This resulted in us not removing the cookie, and when the server would come back online, the app would reload and the user would be logged in.

We should remove `cookie`, just like we remove `currentUserInfo` on logout success.

Test Plan:
Tested that the user is not logged in in the following scenario:
1. login
2. kill the keyserver
3. logout
4. refresh (app not working)
5. run keyserver
6. refresh -> user used to be loogged in, but isn't anymore
Reviewers:
kamil, michal

Subscribers: