[lib] Rename threadInfos to resolvedThreadInfos in getChatMentionCandidates
Summary:
Just a quick rename diff to make it clear that the thread infos passed in are resolved. I'm doing this because the following diffs will involve passing in the unresolved `ThreadInfo`s as a parameter alongside the resolved thread infos.

Part of [[ https://linear.app/comm/issue/ENG-5436/update-usechatmentioncandidatesobjandutils-to-support-ing-chats-by | ENG-5436 ]]

Test Plan: Ran `flow` to confirm no errors

Reviewers: atul, ginsu, inka

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10458