[web] introduce card component
Summary: This component introduces the "Card" component. I built the "Card" component as the new UI is moslty made up of these "Cards" to house various parts of Comm, and I wanted to make sure we could easily enforce the style rules that every "Card" will need. The "Card" component is made up of two parts: the header and the body.

Test Plan: Please see the following screenshots

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10500