[lib] Mark ks as temporary if version or verify_invite_link were called
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6302/update-set-new-session-call-sites | ENG-6302 ]]
If the clients calls an endpoint without authenticating with the keyserver first, they want to temporarily add the keyserver to the store, but not persist it

> If VERSION or VERIFY_INVITE_LINK endpoints were called, we should somehow mark the keyserver as “temporarily” connected - keyserverConnectionHandler will not be attempting to log the user in to that keyserver, and this keyserver will not be persisted.
>1. It might be useful to use the same cookie that we got from the keyserver, in the next request
>2. But calling those endpoints doesn’t necessarily mean that the user wants to permanently connect (login) to this keyserver. The user might for example resign from accepting an invite link.

Note that if the keyserver was already in the store it won't be marked as temporary.

Test Plan:
Tested that when `getVersion` endpoint is hit, the keyserver is added with `connectionIssue: 'temporarily_connected'`.

Reviewers:
ginsu, michal, kamil
Subscribers: