[web] Update `DisconnectedBar` to display based on `useNetworkConnected()`
Summary:
Part of `https://linear.app/comm/issue/ENG-6137/modify-disconnectedbar-so-it-pulls-from-connectivityinfo`.

To begin, I'm going to update the `web` and `native` components to skip eg `useShouldShowDisconnectedBar`, etc. and just get the boolean they need (is the network connected or not) "directly" from `useNetworkConnected` on `web` and `(state => state.connectivity.connected)` on `native`.

I'll then go ahead and rename all of eg `useDisconnectedBarVisibilityHandler` to make it clear that they're no longer connected to the `DisconnectedBar`. I think we'll still want all the Redux actions and whatnot for the "Keyserver status"? page that @ginsu is working on, so I won't rip them out.

Test Plan:
1. Open local `web` app
2. Kill local `keyserver`
3. Observe that nothing happens:
4. Disconnect laptop from WiFi
5. Observe the `DisconnectedBar`:

Reviewers:
ashoat, ginsu, tomek