[web] rename navigation sidebar tooltip to label tooltip
Summary: Rename diff. In D10472, I extended the api of the navigation sidebar tooltip so that this tooltip can be better reused in other places in the web app. Now that this tooltip is more generic, I decided to rename this tooltip into label tooltip since this tooltip just shows a floating label for the elmenet that is hovered over

Test Plan: `flow` and confirmed that this tooltip has no regressions

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10473