[web] cleanup and better organize code for tooltips
Summary: Move diff. As I have been working with the tooltip code recenlty I've noted that the tooltip code is all over the place and is not organized well. I wanted to take a moment to "clean up" the neighborhood and put all the tooltip code into a singular place

Test Plan: `flow` and confirmed that all the tooltips (message action, reaction, and label) had no regressions

Reviewers:

Subscribers: