[web] update colors in app directory modal to use enum option color family
Summary:
As I was working on the web app redesign I took a more critical look at the button color family. I found one variant of the button color family (`--button-background-tetriary-default`) where it was being used in a situation that seemed to better fit the enum option color family. This diff moves that button color family variant into a new enum option color family variant.

For context this is where `--button-background-tetriary-default` was being used:
{F1013593}

To me this is more of an "enum option" element type

Test Plan: Confirmed that the app directory modal which used this color family variant looks exactly the same as before

Reviewers: atul, rohan, kamil

Reviewed By: rohan

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10496