[keyserver] separate olm session creation and persistence
Summary:
separating olm session creation and persistence so that we can create a session before we have a cookie ID with which to persist it.

the next diff shows why this is needed.

abandoning D9345 in favor of this diff since this stack makes it easier to review.

Depends on D10490

Test Plan: tested in last diff in stack by calling the auth responder and confirming that the content olm session is created and stored successfully

Reviewers: ashoat, marcin

Reviewed By: ashoat, marcin

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10491