[lib/web/native] Show viewer in @-mentioning tooltip
Summary:
This enables users to see themselves in the @-mention tooltip so they can mention themselves. We won't have to make any changes to the user-profile on `web`/`native` since we already only just show the 'Message' button if `otherUserInfo` is undefined.

**This does not need to be landed if we decide we don't want to enable this feature**

This is part of [[ https://linear.app/comm/issue/ENG-6297/consider-allowing-users-to-mention-themselves | ENG-6297 ]]

Test Plan:
Please see the videos below

{F1013471}

{F1013475}

Reviewers: atul, ginsu, inka, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10494