[lib][web][native] Check if keyserver is present in store when adding it, unmark temporary
Summary:
issue: [[ https://linear.app/comm/issue/ENG-6097/set-new-session-action-for-new-keyserver-corrupts-keyserverstore | ENG-6097 ]]
We need to remove the temporary flag when a keyserver is being intentionally added
It is possible that the keyserver is already present in the store. So we are only passing the values that we actually want to set, and using the default values or the values from the keyserver store for other fields

Test Plan:
Tested that adding a keyserver still works and doesn't throw errors (tested on web). When we are adding a new keyserver it is already present in the store during `addKeyserverActionType`, because it was added by `setNewSessionActionType` during call to `version` endpoint

Reviewers:
michal, kamil, ginsu
Subscribers: