[web] introduce community header actions
Summary:
This component houses the invite link button and the community settings button, two possible community action buttons that can be rendered, and all the related logic for that. If a community does not have an active invite link then the invite link button should not be rendered.

The invite link button should push the `ViewInviteLinkModal` and the community settings button will navigate the user to the community settings screen (will be done in a subsequent diff)

Linear task: https://linear.app/comm/issue/ENG-5933/top-bar

Depends on D10497

Test Plan:
Please see the screenshots + demo video below

When the community has an active invite link
{F1013673}

When the community does not have an active invite link

{F1013676}

Reviewers: atul, kamil, rohan

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10498