[lib] Reset globalThemeInfo and dataLoaded on identity login
Summary:
issue: [[ https://linear.app/comm/issue/ENG-5823/stop-resetting-globalthemeinfo-on-keyserver-login-do-it-on-identity | ENG-5823 ]]
Test Plan:
Dispatched the action, checked that globalThemeInfo and dataLoaded get updated
Tested with combination of setting currentUserInfo on this action, as is supposed to happen. The reason is that `dataLoaded` is used to hide / show the logged out screen, using this function:
```
const isLoggedIn = (state: BaseAppState<>): boolean =>
  !!(
    state.currentUserInfo &&
    !state.currentUserInfo.anonymous &&
    state.dataLoaded
  );
```
The app looked like this if no keyserver was yet connected:

The app works, it's possible to go to settings ans for example log out. We might want to update the UI some day, but it's imporatant that the user can see that they are logged in if they managed to log in to identity, and be able to log out.
Some discussion on this is in [[ https://linear.app/comm/issue/ENG-4211/review-dataloaded-in-redux | ENG-4211 ]]
Reviewers:
michal, kamil, ginsu
Subscribers: