Implement C++ file to expose backup API
Summary:
This differential implements C++ class that allows to safely trigger backup creation and restoration from any place in C++ using GlobalDBSingleton. The real
application ofthis class will be exposing C++ backup API to Rust in the very next diff.

Test Plan: The same as for the parent differential but instead of directly calling `SQLiteQueryExecutor` call methods of this class.

Reviewers: kamil, michal

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D10509