[RN72][skip-ci] Automatic deletion of all unused ESLint disable directives
Summary:
As of ESLint 8, unused disable directives can be [automatically fixed](https://eslint.org/docs/latest/use/migrate-to-8.0.0#-unused-disable-directives-are-now-fixable). In the next diff we'll introduce `--report-unused-disable-directives` to our ESLint invocations. I'm removing the unused ESLint disable directives in this diff (prior to the introduction of `--report-unused-disable-directives`) to avoid unnecessary CI failures.

Depends on D10541

Test Plan: In combination with the next diff, run `yarn eslint:fix` from the repo root

Reviewers: atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D10542